testing phase

+++ client.py
@@ -5,8 +5,8 @@ import subprocess
 socket = socketio.Client()
 
 socket.connect('http://84db0bdc1324.ngrok.io')
-
-print('Silahkan masukkan token yang didapat dari bot dengan command /token: ')
+print('Halo, silahkan menggunakan bot @proyekTOS_bot dengan command /start')
+print('Silahkan masukkan token yang didapat dari bot dengan command /token: '),
 token = input()
 socket.emit('afterInput', token)
 system = platform.system()
@@ -27,27 +27,39 @@ def sysCheck(data):
 #launch Firefox
 @socket.on('client_launchFirefox')
 def client_launchFirefox(data):
-    subprocess.call(["C:\\Program Files\\Mozilla Firefox\\firefox.exe"])
+    if system.lower() == 'windows':
+        subprocess.call(["C:\\Program Files\\Mozilla Firefox\\firefox.exe"])
+    elif system.lower() == 'linux':
+        subprocess.call(['firefox'])
     msg = 'Firefox telah dijalankan'
     socket.emit('client_firefoxLaunched', {'id' : data['id'], 'room': data['room'], 'message' : msg})
 
 #launch Chrome
 @socket.on('client_launchChrome')
 def client_launchChrome(data):
-    subprocess.call(["C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe"])
-    msg = 'Chrome telah dijalankan'
+    if system.lower() == 'windows':
+        subprocess.call(["C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe"])
+        msg = 'Chrome telah dijalankan'
+    elif system.lower() == 'linux':
+        msg = 'Maaf kami tidak support chrome di Linux'
     socket.emit('client_chromeLaunched', {'id' : data['id'], 'room': data['room'], 'message' : msg})
 
 #shutdown
 @socket.on('client_turnOff')
 def client_turnOff(data):
-    subprocess.call(["shutdown -s"])
+    if system.lower() == 'windows':
+        subprocess.call(["shutdown -s"])
+    elif system.lower() == 'linux':
+        subprocess.call(['shutdown now'])
     msg = 'PC / Laptop anda telah menerima command shutdown'
     socket.emit('client_turnedOff', {'id' : data['id'], 'room': data['room'], 'message' : msg})
 
 #restart
 @socket.on('client_restart')
 def client_restart(data):
-    subprocess.call(["shutdown -r"])
+    if system.lower() == 'windows':
+        subprocess.call(["shutdown -r"])
+    elif system.lower() == 'linux':
+        subprocess.call(['reboot'])
     msg = 'PC / Laptop anda telah menerima command restart'
     socket.emit('client_restarted', {'id' : data['id'], 'room': data['room'], 'message' : msg})
